+++ gendarme/rules/Gendarme.Rules.Exceptions/InstantiateArgumentExceptionCorrectlyRule.cs
@@ -189,8 +189,13 @@ namespace Gendarme.Rules.Exceptions
       // don't process methods without parameters unless it's a special method (e.g. a property)
       // this cover cases like "if (x == null) CallLocalizedThrow();" and the inner type compilers
       // generates for yield/iterator (a field is used)
-      if (!method.IsSpecialName && !method.HasParameters)
-        return RuleResult.DoesNotApply;
+      // if (!method.IsSpecialName && !method.HasParameters)
+      //   return RuleResult.DoesNotApply;
+
+      // A possible resolution for `<TypesInside>d__1.MoveNext()` and the like
+      // Still softens the FxCop rule behaviour, though
+      // if (method.DeclaringType.HasCompilerGeneratedAttribute() && !method.HasParameters)
+      //    return RuleResult.DoesNotApply;
 
       // and when the IL contains a NewObj instruction
       if (!OpCodeEngine.GetBitmask(method).Get(Code.Newobj))